+++ ocd/packages/desktop/src/components/OcdCanvas.tsx
@@ -39,7 +39,7 @@ export const OcdCanvas = ({ dragData, setDragData, ocdConsoleConfig, ocdDocument
     const {activeFile, setActiveFile} = useContext(ActiveFileContext)
     const uuid = () => `gid-${uuidv4()}`
     const page: OcdViewPage = ocdDocument.getActivePage()
-    const allCompartmentIds = ocdDocument.getOciResourceList('comparment').map((r) => r.id)
+    // const allCompartmentIds = ocdDocument.getOciResourceList('comparment').map((r) => r.id)
     const visibleLayers = page.layers.filter((l: OcdViewLayer) => l.visible).map((l: OcdViewLayer) => l.id)
     // const visibleResourceIds = ocdDocument.getResources().filter((r: OcdResource) => visibleLayers.includes(r.compartmentId) || (!allCompartmentIds.includes(r.compartmentId) && r.resourceType !== 'Compartment')).map((r: any) => r.id)
     const visibleResourceIds = ocdDocument.getResources().filter((r: any) => visibleLayers.includes(r.compartmentId)).map((r: any) => r.id)
@@ -172,7 +172,8 @@ export const OcdCanvas = ({ dragData, setDragData, ocdConsoleConfig, ocdDocument
             // Redraw
             console.info('OcdCanvas: Design:', ocdDocument.design)
             setOcdDocument(OcdDocument.clone(ocdDocument))
-            if (!activeFile.modified) setActiveFile({name: activeFile.name, modified: true})
+            if (!activeFile.modified) activeFile.modified = true
+            // if (!activeFile.modified) setActiveFile({name: activeFile.name, modified: true})
         }
         return false
     }
@@ -407,6 +408,7 @@ export const OcdCanvas = ({ dragData, setDragData, ocdConsoleConfig, ocdDocument
             onDrop={(e) => onDrop(e)}
             onDragLeave={(e) => onDragLeave()}
             onDragOver={(e) => onDragOver(e)}
+            aria-hidden
             >
             <svg className='ocd-designer-canvas-svg'
                 id='canvas_root_svg'