+++ ocd/packages/desktop/src/components/properties/OcdPropertyTypes.tsx
@@ -128,7 +128,7 @@ export const isPropertyDisplayConditionTrue = (conditional: boolean, condition: 
 
 export const OcdDisplayNameProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue] = useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -140,7 +140,7 @@ export const OcdDisplayNameProperty = ({ ocdDocument, setOcdDocument, resource, 
         e.target.reportValidity()
         resource[attribute.key] = e.target.value
         setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdDisplayNameProperty: ${attribute.id} Render(${value})`)
@@ -154,7 +154,7 @@ export const OcdDisplayNameProperty = ({ ocdDocument, setOcdDocument, resource, 
 
 export const OcdTextProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -166,7 +166,7 @@ export const OcdTextProperty = ({ ocdDocument, setOcdDocument, resource, config,
         e.target.reportValidity()
         resource[attribute.key] = e.target.value
         // setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdTextProperty: ${attribute.id} Render(${value})`)
@@ -180,7 +180,7 @@ export const OcdTextProperty = ({ ocdDocument, setOcdDocument, resource, config,
 
 export const OcdNumberProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -192,7 +192,7 @@ export const OcdNumberProperty = ({ ocdDocument, setOcdDocument, resource, confi
         e.target.reportValidity()
         resource[attribute.key] = e.target.value
         // setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdNumberProperty: ${attribute.id} Render(${value})`)
@@ -206,7 +206,7 @@ export const OcdNumberProperty = ({ ocdDocument, setOcdDocument, resource, confi
 
 export const OcdBooleanProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -214,7 +214,7 @@ export const OcdBooleanProperty = ({ ocdDocument, setOcdDocument, resource, conf
         setValue(e.target.checked)
         resource[attribute.key] = e.target.checked
         // setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdBooleanProperty: ${attribute.id} Render(${value})`)
@@ -228,7 +228,7 @@ export const OcdBooleanProperty = ({ ocdDocument, setOcdDocument, resource, conf
 
 export const OcdCodeProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     const onChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
@@ -240,7 +240,7 @@ export const OcdCodeProperty = ({ ocdDocument, setOcdDocument, resource, config,
         e.target.reportValidity()
         resource[attribute.key] = e.target.value
         // setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdCodeProperty: ${attribute.id} Render(${value})`)
@@ -270,7 +270,7 @@ export const OcdLookupGroupOption = ({group}: {group: ResourceElementConfigLooku
 export const OcdLookupProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     console.debug('OcdPropertyTypes: OcdLookupProperty', config, attribute, resource)
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     const lookupGroups = config && config.lookupGroups ? config.lookupGroups : []
@@ -288,7 +288,7 @@ export const OcdLookupProperty = ({ ocdDocument, setOcdDocument, resource, confi
         setValue(e.target.value)
         // TODO: Replace with a performant Solution
         setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdLookupProperty: ${attribute.id} Render(${value})`)
@@ -311,7 +311,7 @@ export const OcdLookupProperty = ({ ocdDocument, setOcdDocument, resource, confi
 }
 
 export const OcdLookupListProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     const resourceType = OcdUtils.toResourceType(attribute.lookupResource)
@@ -326,7 +326,7 @@ export const OcdLookupListProperty = ({ ocdDocument, setOcdDocument, resource, c
         else resource[attribute.key] = resource[attribute.key].filter((s: string) => s !== id)
         // TODO: Replace with a performant Solution
         setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdLookupListProperty: ${attribute.id} Render(${value})`)
@@ -346,7 +346,7 @@ export const OcdLookupListProperty = ({ ocdDocument, setOcdDocument, resource, c
 
 export const OcdStaticLookupProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue] = useState(resource[attribute.key])
     const properties = config?.properties ? config.properties : {}
     const configOptions = config?.options ? config.options : []
@@ -360,7 +360,7 @@ export const OcdStaticLookupProperty = ({ ocdDocument, setOcdDocument, resource,
         setValue(e.target.value)
         // TODO: Replace with a performant Solution
         setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     useEffect(() => {
         if (!resource[attribute.key] || resource[attribute.key] === '') {
@@ -368,7 +368,7 @@ export const OcdStaticLookupProperty = ({ ocdDocument, setOcdDocument, resource,
                 resource[attribute.key] = resources[0].id
                 // TODO: Replace with a performant Solution
                 // setOcdDocument(OcdDocument.clone(ocdDocument))
-                if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+                if (!activeFile.modified) activeFile.modified = true
             }
         }
     }, [])
@@ -389,8 +389,8 @@ export const OcdStaticLookupProperty = ({ ocdDocument, setOcdDocument, resource,
 
 export const OcdCacheLookupProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     console.debug('OcdPropertyTypes: OcdCacheLookupProperty', config, attribute, resource)
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
-    const {ocdCache, setOcdCache} = useContext(CacheContext)
+    const {activeFile} = useContext(ActiveFileContext)
+    const {ocdCache} = useContext(CacheContext)
     const [value, setValue]= useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     // const id = `${rootResource.id.replace(/\W+/g, "")}_${attribute.id.replace(/\W+/g, "")}`
@@ -414,7 +414,7 @@ export const OcdCacheLookupProperty = ({ ocdDocument, setOcdDocument, resource, 
         setValue(e.target.value)
         // TODO: Replace with a performant Solution
         setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug('OcdPropertyTypes: OcdCacheLookupProperty', config, attribute, resource, resources)
@@ -437,7 +437,7 @@ export const OcdCacheLookupProperty = ({ ocdDocument, setOcdDocument, resource, 
 
 export const OcdStringListProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key] ? resource[attribute.key].join(',') : '')
     const properties = config && config.properties ? config.properties : {}
     const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -449,7 +449,7 @@ export const OcdStringListProperty = ({ ocdDocument, setOcdDocument, resource, c
         e.target.reportValidity()
         resource[attribute.key] = e.target.value.split(',').filter((v) => v !== '')
         // setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdStringListProperty: ${attribute.id} Render(${value})`)
@@ -463,7 +463,7 @@ export const OcdStringListProperty = ({ ocdDocument, setOcdDocument, resource, c
 
 export const OcdNumberListProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
     const id = useId()
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key].join(','))
     const properties = config && config.properties ? config.properties : {}
     const onChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -475,7 +475,7 @@ export const OcdNumberListProperty = ({ ocdDocument, setOcdDocument, resource, c
         e.target.reportValidity()
         resource[attribute.key] = e.target.value.split(',')
         // setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdNumberListProperty: ${attribute.id} Render(${value})`)
@@ -488,7 +488,7 @@ export const OcdNumberListProperty = ({ ocdDocument, setOcdDocument, resource, c
 }
 
 export const OcdListProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     // const id = `${rootResource.id.replace(/\W+/g, "")}_${attribute.id.replace(/\W+/g, "")}`
@@ -502,7 +502,7 @@ export const OcdListProperty = ({ ocdDocument, setOcdDocument, resource, config,
 }
 
 export const OcdSetProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     // const id = `${rootResource.id.replace(/\W+/g, "")}_${attribute.id.replace(/\W+/g, "")}`
@@ -516,7 +516,7 @@ export const OcdSetProperty = ({ ocdDocument, setOcdDocument, resource, config, 
 }
 
 export const OcdSetLookupProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const properties = config && config.properties ? config.properties : {}
     const resourceType = OcdUtils.toResourceType(attribute.lookupResource)
@@ -529,7 +529,7 @@ export const OcdSetLookupProperty = ({ ocdDocument, setOcdDocument, resource, co
         if (checked) resource[attribute.key].push(securityListId)
         else resource[attribute.key] = resource[attribute.key].filter((s: string) => s !== securityListId)
         // setOcdDocument(OcdDocument.clone(ocdDocument))
-        if(!activeFile.modified) setActiveFile({...activeFile, modified: true})
+        if (!activeFile.modified) activeFile.modified = true
     }
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdSetLookupProperty: ${attribute.id} Render(${value})`)
@@ -548,7 +548,7 @@ export const OcdSetLookupProperty = ({ ocdDocument, setOcdDocument, resource, co
 }
 
 export const OcdMapProperty = ({ ocdDocument, setOcdDocument, resource, config, attribute, rootResource }: ResourceProperty): JSX.Element => {
-    const {activeFile, setActiveFile} = useContext(ActiveFileContext)
+    const {activeFile} = useContext(ActiveFileContext)
     const [value, setValue]= useState(resource[attribute.key])
     const className = isPropertyDisplayConditionTrue(attribute.conditional, attribute.condition, resource, rootResource) ? `ocd-property-row ocd-simple-property-row` : `collapsed hidden`
     console.debug(`>>>> OcdPropertyTypes: OcdMapProperty: ${attribute.id} Render(${value})`)