+++ okitweb/static/okit/model/js/resources/node_pool.js
@@ -48,7 +48,15 @@ class NodePool extends OkitArtifact {
         Object.defineProperty(this, 'node_pool_type', {get: function() {return this.getOkitJson().getResource(this.cluster_id).node_pool_type;}});
         Object.defineProperty(this, 'virtual_nodes', {get: function() {return this.node_pool_type === 'Virtual'}});
     }
-
+    /*
+    ** Conversion Routine allowing loading of old json
+     */
+    convert() {
+        super.convert();
+        this.node_config_details.placement_configs.forEach((p) => p.availability_domain = String(p.availability_domain).slice(-1))
+        if (this.nodes) delete this.nodes
+    }
+    
     isVirtualNodes = () => this.getOkitJson().getResource(this.cluster_id) && this.getOkitJson().getResource(this.cluster_id).node_pool_type === 'Virtual'
     setDefaultShape = () => this.node_shape = this.isVirtualNodes() ? 'Pod.Standard.E3.Flex' : 'VM.Standard.E3.Flex'